% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclers.R
\name{cycler_normalise}
\alias{cycler_normalise}
\title{Normalise a string to represent a cycler name}
\usage{
cycler_normalise(cycler_type)
}
\arguments{
\item{cycler_type}{The string to normaliase}
}
\value{
The normalised string
}
\description{
Normalise a string to represent a cycler name
}
