% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclers.R
\name{cycler_types}
\alias{cycler_types}
\alias{supported_file_extensions}
\title{List of supported cyclers and their file extensions}
\usage{
cycler_types()

supported_file_extensions()
}
\description{
List of supported cyclers and their file extensions
}
