% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echem.Data.R
\name{echem.data.object}
\alias{echem.data.object}
\alias{is.echem.data.object}
\alias{as.echem.data.object}
\title{Echem data objects}
\usage{
echem.data.object(...)

is.echem.data.object(x)

as.echem.data.object(...)
}
\arguments{
\item{x}{The object}
}
\description{
Echem data objects
}
\keyword{internal}
