% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arbin.R
\name{.get_arbin_attributes}
\alias{.get_arbin_attributes}
\title{Get additional attributes from an arbin excel file}
\usage{
.get_arbin_attributes(file)
}
\arguments{
\item{file}{Path to the arbin file}
}
\value{
A list of attributes that were found
}
\description{
This function finds the date, channel and schedule name for an arbin dataset
}
\keyword{internal}
