% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biologic.R
\name{.get_biologic_attributes}
\alias{.get_biologic_attributes}
\title{Get additional attributes from a biologic mpt file}
\usage{
.get_biologic_attributes(header_text)
}
\arguments{
\item{header_lines}{The header of a biologic mpt file}
}
\value{
A list of attributes that were found
}
\description{
This function finds the date and channel for a biologic dataset
}
\keyword{internal}
