% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{.get_internal_attributes}
\alias{.get_internal_attributes}
\title{Get additional attributes from an exported csv file}
\usage{
.get_internal_attributes(header_text)
}
\arguments{
\item{header_lines}{The header of an exported csv file}
}
\value{
A list of attributes that were found
}
\description{
This function finds the attributes for an dataset
}
\keyword{internal}
