% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclers.R
\name{guess_cycler}
\alias{guess_cycler}
\title{Guess cycler for file}
\usage{
guess_cycler(file)
}
\arguments{
\item{file}{Path to data file}
}
\value{
The cycler type
}
\description{
A very basic function to choose a cycler based on the file extension, implementation may change if more cyclers are added
}
