% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arbin.R
\name{load.arbin}
\alias{load.arbin}
\title{Read Echem Data for Arbin Cycler}
\usage{
load.arbin(file)
}
\arguments{
\item{file}{Path to data file}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data from an Arbin xls file.
}
\examples{
load.arbin("/path/to/file.xls")
}
\keyword{internal}
