% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biologic.R
\name{load.biologic.mpr}
\alias{load.biologic.mpr}
\title{Read Echem Data for Biologic Cycler}
\usage{
load.biologic.mpr(mprfile)
}
\arguments{
\item{file}{Path to data file}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data from a biologic mpr file.
Originally based on Chris Kerr's python script (https://github.com/chatcannon/galvani) and modified to support additional data types.
}
\examples{
load.biologic.mpr("/path/to/file.mpr")
}
\keyword{internal}
