% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biologic.R
\name{load.biologic.mpt}
\alias{load.biologic.mpt}
\title{Read Echem Data for Biologic Cycler}
\usage{
load.biologic.mpt(file)
}
\arguments{
\item{file}{Path to data file}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data from a biologic mpt file.
}
\examples{
load.biologic.mpt("/path/to/file.mpt")
}
\keyword{internal}
