% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biologic.R
\name{load.biologic}
\alias{load.biologic}
\title{Read Echem Data for Biologic Cycler}
\usage{
load.biologic(file)
}
\arguments{
\item{file}{Path to data file}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data from a biologic mpt or mpr file.
}
\examples{
load.biologic("/path/to/file.mpt")
load.biologic("/path/to/file.mpr")
}
\keyword{internal}
