% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R
\name{load.internal}
\alias{load.internal}
\title{Read echem data that has previously been exported to a csv file}
\usage{
load.internal(file)
}
\arguments{
\item{file}{Path to data file}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data from a csv file that was generated using \code{\link{export_echem}}.
}
\examples{
load.internal("/path/to/file.csv")
}
\keyword{internal}
