% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ivium.R
\name{load.ivium}
\alias{load.ivium}
\title{Read Echem Data for ivium Cycler}
\usage{
load.ivium(file)
}
\arguments{
\item{file}{Path to data file}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data from an ivium idf or txt file.
}
\examples{
load.ivium("/path/to/file.txt")
}
\keyword{internal}
