% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Land.R
\name{load.land}
\alias{load.land}
\title{Read Echem Data for land (lahne) Cycler}
\usage{
load.land(file)
}
\arguments{
\item{file}{Path to data file}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data from a land txt file.
}
\examples{
load.land("/path/to/file.txt")
}
\keyword{internal}
