% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maccor.R
\name{load.maccor}
\alias{load.maccor}
\title{Read Echem Data for maccor Cycler}
\usage{
load.maccor(file)
}
\arguments{
\item{file}{Path to data file}
}
\value{
A data frame containing the echem data
}
\description{
This function is used to import echem data from a maccor txt file.
}
\examples{
load.maccor("/path/to/file.txt")
}
\keyword{internal}
