% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.echem.R
\name{plot.echem.data.object}
\alias{plot.echem.data.object}
\title{Plot echem data}
\usage{
\method{plot}{echem.data.object}(x, ..., frac = c(F, T), div = c(5, 2))
}
\arguments{
\item{x}{An echem.data.object}

\item{...}{Additional parameters passed to \code{\link[jms.classes]{plot.jms.data.object}}}
}
\description{
This function is used to plot echem data.
}
