% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.echem.R
\name{read.echem}
\alias{read.echem}
\title{Read Echem Data}
\usage{
read.echem(path, ...)
}
\arguments{
\item{echem_file}{Path to data file}
}
\value{
An echem.data.object containing the echem data
}
\description{
This function is used to import echem data.
}
\examples{
read.echem("/path/to/file.mpr")
read.echem("/path/to/file.mpt")
read.echem("/path/to/file.xlsx")
read.echem("/path/to/file.txt")
read.echem("/path/to/file.idf")
}
